Fix X Crash on Quit & Restart
This fixes the X server crashes I was getting when recompiling &
restarting xmonad. I thought that the dynamicEventHook was getting
called on shutdown to kill the xmobar threads, but that only happens on
screen count changes.

So this adds a recompileAndRestart X action that recompiles the
configuration, and then shuts down the status bars and restarts xmonad
if the compilation was successful.

It now shuts down the status bars when exiting as well.