Update Stack LTS & Xmobar
Update Xmobar to the latest upstream commit. This allows us to use the
Xmobar-as-a-library code that I developed, which greatly simplifies the
XmobarStub module since we no longer have to re-implement the xmobar
exectuable's code.

This requires some import restructuring and updating the Stack LTS to
pull in a compatible version of the libmpd library.

The next step of simplifying the xmobar integration is adding a new
Handle or TBQueue Reader module so that we do not have to use external
named pipes to communicate between XMonad and Xmobar.