Use Handles to Communicate With Xmobar
Use my new HandleReader Xmobar plugin to send the workspace list, layout
type, & window list to Xmobar - instead of named pipes.

This drastically reduces the code needed in the XmobarStub module,
although we will probably move the StatusBarStorage type & functions to
this module so that we can contain the code to start/stop the xmobar
processes and open/close the Handles. That code may be added to
xmonad-contrib or a separate package.