Refactor Xmobar Process Tracking & Cleanup
Move the StatusBarStorage type to the XmobarStub module and expand the
stored data to include the input & output handles.

Move the terminateProcesses function to the XmobarStub module and have
it close the input & output handles in addition to stopping the Xmobar
processes.

Make the XmobarStub.run function responsible for adding the new Xmobar
instance's data to the StatusBarStorage, simplifying the runXmobar
function in the StatusBar module.